db: Move enterprise/frontend/db to enterprise/internal/db
